move other firmware tests to new file
